+++ plugins/chat/assets/javascripts/discourse/components/chat-channel-card.hbs
@@ -10,7 +10,7 @@
   >
     <div class="chat-channel-card__header">
       <LinkTo
-        @route="chat.channel.from-params"
+        @route="chat.channel"
         @models={{@channel.routeModels}}
         class="chat-channel-card__name-container"
       >
