+++ plugins/chat/assets/javascripts/discourse/components/chat-channel-row.hbs
@@ -1,5 +1,5 @@
 <LinkTo
-  @route="chat.channel.from-params"
+  @route="chat.channel"
   @models={{@channel.routeModels}}
   class={{concat-class
     "chat-channel-row"
