+++ plugins/chat/assets/javascripts/discourse/components/chat-drawer.js
@@ -236,7 +236,7 @@ export default Component.extend({
         this.set("view", DRAFT_CHANNEL_VIEW);
         this.appEvents.trigger("chat:float-toggled", false);
         return;
-      case "chat.channel.from-params":
+      case "chat.channel":
         return this._openChannel(
           route.parent.params.channelId,
           this._highlightCb(route.queryParams.messageId)
