+++ plugins/chat/assets/javascripts/discourse/components/chat-live-pane.hbs
@@ -15,7 +15,7 @@
       {{/if}}
 
       <LinkTo
-        @route={{this.infoTabRoute}}
+        @route="chat.channel.info"
         @models={{this.chatChannel.routeModels}}
         class="chat-channel-title-wrapper"
       >
