+++ plugins/chat/assets/javascripts/discourse/components/full-page-chat.hbs
@@ -3,5 +3,6 @@
     @chatChannel={{this.chat.activeChannel}}
     @onBackClick={{action "navigateToIndex"}}
     @onSwitchChannel={{action "switchChannel"}}
+    @targetMessageId={{readonly @targetMessageId}}
   />
 {{/if}}
