+++ plugins/chat/assets/javascripts/discourse/controllers/chat-channel.js
@@ -5,6 +5,8 @@ import { inject as service } from "@ember/service";
 export default class ChatChannelController extends Controller {
   @service chat;
 
+  targetMessageId = null;
+
   // Backwards-compatibility
   queryParams = ["messageId"];
