+++ plugins/chat/assets/javascripts/discourse/routes/chat-channel-info.js
@@ -1,7 +1,9 @@
 import DiscourseRoute from "discourse/routes/discourse";
 import { inject as service } from "@ember/service";
 import { ORIGINS } from "discourse/plugins/chat/discourse/services/chat-channel-info-route-origin-manager";
+import withChatChannel from "./chat-channel-decorator";
 
+@withChatChannel
 export default class ChatChannelInfoRoute extends DiscourseRoute {
   @service chatChannelInfoRouteOriginManager;
