+++ plugins/chat/assets/javascripts/discourse/routes/chat-channel-legacy.js
@@ -9,13 +9,8 @@ export default class ChatChannelLegacyRoute extends DiscourseRoute {
       this.routeName
     );
 
-    this.router.replaceWith(
-      "chat.channel.from-params",
-      channelTitle,
-      channelId,
-      {
-        queryParams: { messageId },
-      }
-    );
+    this.router.replaceWith("chat.channel", channelTitle, channelId, {
+      queryParams: { messageId },
+    });
   }
 }
