+++ plugins/chat/assets/javascripts/discourse/routes/chat.js
@@ -22,7 +22,6 @@ export default class ChatRoute extends DiscourseRoute {
 
     const INTERCEPTABLE_ROUTES = [
       "chat.channel",
-      "chat.channel.from-params",
       "chat.channel.near-message",
       "chat.channel-legacy",
       "chat",
