--- plugins/chat/assets/javascripts/discourse/templates/chat-channel-from-params.hbs
@@ -1,1 +1,0 @@
-<FullPageChat />
