+++ plugins/chat/assets/javascripts/discourse/templates/chat-channel-info.hbs
@@ -12,7 +12,7 @@
           </LinkTo>
         {{else}}
           <LinkTo
-            @route="chat.channel.from-params"
+            @route="chat.channel"
             @models={{this.model.routeModels}}
             class="chat-full-page-header__back-btn no-text btn-flat btn"
             title={{i18n "chat.channel_info.back_to_channel"}}
