+++ plugins/chat/assets/javascripts/discourse/templates/chat-channel.hbs
@@ -1,1 +1,1 @@
-{{outlet}}
+<FullPageChat @targetMessageId={{this.targetMessageId}} />
